conn_test.go: Fold ICE state test into conn_test.go
