+++ conn_test.go
@@ -5,6 +5,8 @@ import (
 	"errors"
 	"net"
 	"testing"
+
+	"github.com/pion/webrtc/v4"
 )
 
 func TestClosedWriteError(t *testing.T) {
@@ -59,5 +61,22 @@ func TestConnReadKeepsRemainderWhenBufferIsShort(t *testing.T) {
 	}
 	if got := string(b[:n]); got != "o" {
 		t.Fatalf("third Read() = %q, want %q", got, "o")
+	}
+}
+
+func TestIsTerminalICEState(t *testing.T) {
+	for state, terminal := range map[webrtc.ICETransportState]bool{
+		webrtc.ICETransportStateUnknown:      false,
+		webrtc.ICETransportStateNew:          false,
+		webrtc.ICETransportStateChecking:     false,
+		webrtc.ICETransportStateConnected:    false,
+		webrtc.ICETransportStateCompleted:    false,
+		webrtc.ICETransportStateDisconnected: false,
+		webrtc.ICETransportStateFailed:       true,
+		webrtc.ICETransportStateClosed:       true,
+	} {
+		if got := isTerminalICEState(state); got != terminal {
+			t.Errorf("isTerminalICEState(%s) = %t, want %t", state, got, terminal)
+		}
 	}
 }
